更新认证权限缓存

+++ src/main/java/cn/rtt/server/system/domain/request/LoginRequest.java
@@ -34,6 +34,6 @@ public class LoginRequest
     /**
      * 滑动验证
      */
-    private Double slideOffset;
+    private Long slideOffset;
 
 }
